Remove delay-file-viewer flag and remove async file extraction from backend.
The frontend still supports it, I didn't want to touch untested code
tbh.

Fixes #4203, refs #3579
